+++ backend/src/baserow/contrib/database/views/models.py
@@ -88,6 +88,10 @@ class View(
         blank=True,
         help_text="The password required to access the public view URL.",
     )
+    show_logo = models.BooleanField(
+        default=True,
+        help_text="Indicates whether the logo should be shown in the public view.",
+    )
 
     @property
     def public_view_has_password(self) -> bool: