+++ premium/web-frontend/modules/baserow_premium/plugins.js
@@ -1,5 +1,6 @@
 import { BaserowPlugin } from '@baserow/modules/core/plugins'
 import PremiumTopSidebar from '@baserow_premium/components/sidebar/PremiumTopSidebar'
+import BaserowLogoShareLinkOption from '@baserow_premium/components/views/BaserowLogoShareLinkOption'
 
 export class PremiumPlugin extends BaserowPlugin {
   static getType() {
@@ -26,5 +27,9 @@ export class PremiumPlugin extends BaserowPlugin {
 
   getSidebarTopComponent() {
     return PremiumTopSidebar
+  }
+
+  getAdditionalShareLinkOptions() {
+    return [BaserowLogoShareLinkOption]
   }
 }