+++ web-frontend/modules/core/components/BaserowLogo.vue
@@ -1,0 +1,19 @@
+<template>
+  <a
+    href="https://baserow.io"
+    target="_blank"
+    title="Baserow - open source no-code database tool and Airtable alternative"
+  >
+    <img
+      height="14"
+      src="@baserow/modules/core/static/img/logo.svg"
+      alt="Baserow - open source no-code database tool and Airtable alternative"
+    />
+  </a>
+</template>
+
+<script>
+export default {
+  name: 'BaserowLogo',
+}
+</script>