+++ web-frontend/modules/core/plugins.js
@@ -82,4 +82,12 @@ export class BaserowPlugin extends Registerable {
   getAppLayoutComponent() {
     return null
   }
+
+  /**
+   * Every registered plugin can display multiple additional public share link options
+   * which will be visible on the share public view context.
+   */
+  getAdditionalShareLinkOptions() {
+    return []
+  }
 }