+++ web-frontend/modules/database/components/view/form/FormViewModePreviewForm.vue
@@ -104,8 +104,11 @@
         >
         </FormViewField>
       </div>
-      <div class="form-view__actions">
-        <FormViewPoweredBy></FormViewPoweredBy>
+      <div
+        class="form-view__actions"
+        :class="{ 'form-view__actions--single': !view.show_logo }"
+      >
+        <FormViewPoweredBy v-if="view.show_logo"></FormViewPoweredBy>
         <div class="form-view__submit">
           <Editable
             ref="submit_text"