+++ web-frontend/modules/database/components/view/gallery/GalleryViewHeader.vue
@@ -31,7 +31,15 @@
         @update-cover-image-field="updateCoverImageField"
       ></ViewFieldsContext>
     </li>
-    <li class="header__filter-item header__filter-item--right">
+    <li
+      class="
+        header__filter-item header__filter-item--right
+        margin-top-1 margin-right-1
+      "
+    >
+      <BaserowLogo v-if="view.show_logo && isPublic" />
+    </li>
+    <li class="header__filter-item">
       <ViewSearch
         :view="view"
         :fields="fields"
@@ -49,10 +57,11 @@ import { mapGetters, mapState } from 'vuex'
 import { notifyIf } from '@baserow/modules/core/utils/error'
 import ViewFieldsContext from '@baserow/modules/database/components/view/ViewFieldsContext'
 import ViewSearch from '@baserow/modules/database/components/view/ViewSearch'
+import BaserowLogo from '@baserow/modules/core/components/BaserowLogo'
 
 export default {
   name: 'GalleryViewHeader',
-  components: { ViewFieldsContext, ViewSearch },
+  components: { BaserowLogo, ViewFieldsContext, ViewSearch },
   props: {
     database: {
       type: Object,
@@ -82,6 +91,9 @@ export default {
   computed: {
     ...mapState({
       tableLoading: (state) => state.table.loading,
+    }),
+    ...mapGetters({
+      isPublic: 'page/view/public/getIsPublic',
     }),
   },
   beforeCreate() {