+++ web-frontend/modules/database/components/view/grid/GridView.vue
@@ -92,21 +92,6 @@
       @edit-modal="openRowEditModal($event.id)"
       @scroll="scroll($event.pixelY, $event.pixelX)"
     >
-      <template #foot>
-        <div v-if="publicGrid" class="grid-view__foot-logo">
-          <a
-            href="https://baserow.io"
-            target="_blank"
-            title="Baserow - open source no-code database tool and Airtable alternative"
-          >
-            <img
-              height="14"
-              src="@baserow/modules/core/static/img/logo.svg"
-              alt="Baserow - open source no-code database tool and Airtable alternative"
-            />
-          </a>
-        </div>
-      </template>
     </GridViewSection>
     <GridViewRowDragging
       ref="rowDragging"